First 2 days of the "12 Projects Of Codemas"
Bought a Raspberry Pi Pico advent calendar by Pi Hut, the "12 Projects Of Codemas" - a box full of components and a Pi Pico H. I thought it would be a fun way to start learning the Pico and further my knowledge of microelectronics as I've been doing with the Micro:bit and Arduino UNO R3.

Completed Days 1 & 2 of the calendar so far - basic onboard and external LED programs. Looking forward to the other 10! Not following the days strictly (did the first two in one day) but still a fun idea to follow along.