Day 6 - Intro To Light Sensors
Day 6 of the "12 Projects Of Codemas" Pi Pico kit - today's task was an introduction to the use of a photo transistor - the technical term for a type of light sensor. Followed the lesson and made a setup that activated a number of LED lights based on the light reading and percentage.